refactor(editor): compose writing surface 분리
